@ashthespy 2020-05-11
@alcore Sorry I seem to have missed this! Thanks for the effort, and appreciate the benchmarking!
Please go ahead with your `SpotifyObjectId` and `SpotifyContext` as well, we can handle the api changes depending on your implementation..