@sashahilton00 2020-07-24
Just triggering a rebuild. The API changes look fine to me as well. @alcore are you planning on refactoring `SpotifyObjectId` and `SpotifyContext` as well?