@Johannesd3 2021-01-25
I don't think the change is breaking, the function signatures are the same. It's a pity that this isn't merged, the code seems well-written and well-documented... If @alcore isn't interested anymore, I will perhaps create a new pr based on this one (if it's ok).